+++ src/main/java/com/flab/shoeauction/exception/GlobalExceptionHandler.java
@@ -25,7 +25,6 @@ import com.flab.shoeauction.exception.product.ProductNotFoundException;
 import com.flab.shoeauction.exception.user.AuthenticationNumberMismatchException;
 import com.flab.shoeauction.exception.user.DuplicateEmailException;
 import com.flab.shoeauction.exception.user.DuplicateNicknameException;
-import com.flab.shoeauction.exception.user.EmptyCartException;
 import com.flab.shoeauction.exception.user.NotAuthorizedException;
 import com.flab.shoeauction.exception.user.TokenExpiredException;
 import com.flab.shoeauction.exception.user.UnableToChangeNicknameException;
@@ -160,12 +159,6 @@ public class GlobalExceptionHandler {
         ImageRoadFailedException ex) {
         log.debug("이미지 로드에 실패하였습니다.", ex);
         return IMAGE_ROAD_FAILED;
-    }
-
-    @ExceptionHandler(EmptyCartException.class)
-    public final ResponseEntity<Void> handleEmptyCartException (EmptyCartException ex) {
-        log.debug("위시리스트가 비어있습니다.", ex);
-        return BAD_REQUEST;
     }
 
 }
