--- src/main/java/com/flab/shoeauction/exception/user/EmptyCartException.java
@@ -1,8 +1,0 @@
-package com.flab.shoeauction.exception.user;
-
-public class EmptyCartException extends RuntimeException {
-
-    public EmptyCartException(String message) {
-        super(message);
-    }
-}
